refs #1248 initial work on the statement reconnect issue
