+++ src/dbmodulewrap.h
@@ -1,3 +1,5 @@
+/* -*- mode: c++; indent-tabs-mode: nil -*- */
+
 #ifndef SYBASE_SRC_DBMODULEWRAP_H
 #define SYBASE_SRC_DBMODULEWRAP_H
 
@@ -34,7 +36,6 @@ class DBModuleWrap {
         Datasource * ds() { return stmt->getDatasource(); }
         const QoreEncoding * enc() { return ds()->getQoreEncoding(); }
 
-
         QoreString * encodeSQL(const QoreString &s, ExceptionSink* xsink) {
             return s.convertEncoding(enc(), xsink);
         }
@@ -47,8 +48,8 @@ class DBModuleWrap {
         SQLStatement* stmt;
     };
 
-
-    class ModuleWrap {
+public:
+   class ModuleWrap {
         ~ModuleWrap() {}
     public:
         ModuleWrap() : m(new Module), params(0) {}
@@ -102,7 +103,6 @@ class DBModuleWrap {
         }
     }
 
-
     static ModuleWrap * module_wrap(SQLStatement* stmt,
             bool create = true)
     {
@@ -111,6 +111,8 @@ class DBModuleWrap {
             if (!create) return 0;
             mv = new ModuleWrap;
             stmt->setPrivateData(mv);
+            StatementHelper sh(stmt);
+            sh.conn()->registerStatement(mv);
         }
         return mv;
     }
@@ -159,7 +161,6 @@ class DBModuleWrap {
         ModuleWrap::Delete(m, xsink);
         return 0;
     }
-
 
     static int bind(SQLStatement* stmt, const QoreListNode &l,
             ExceptionSink* xsink)
@@ -254,12 +255,11 @@ class DBModuleWrap {
         return run<QoreHashNode *>(&Module::get_output_rows, stmt, xsink);
     }
 
-
     qore_dbi_method_list &methods;
 
   public:
-    DBModuleWrap(qore_dbi_method_list &methods) :
-        methods(methods)
+    DBModuleWrap(qore_dbi_method_list &n_methods) :
+        methods(n_methods)
     {}
 
     void reg() {
